requirements.txt
pydub
requests
python-dotenv
yt-dlp
mutagen
isodate